[smalloc] Refined test_alloc_two_blocks
Found a critical hole in the design of allocator: we can't turn a free
block into busy one because we need a size of previous block but it's
set to 0 for free blocks.

I'll redesign an allocator to use least significant bit of size to
indicate free/busy block. That will decrease size of maximal allocation
in 4 times (it's possible to make only 2-time decrease, but that's
harder to implement and I'm not sure it worth it).